Moved `PubVis` type and `Merge` trait to `utils` module
